feat : 게시글 작성 관련 공통 컴포넌트 이동 & style : 스텝 헤더 뒤로가기 버튼 수정
