+++ frontend/src/components/SelectedSummary.tsx
